Restructured client routes
