+++ app/routes/authenticated/partner/clients/index.js
@@ -1,0 +1,7 @@
+import Route from '@ember/routing/route';
+
+export default class AuthenticatedPartnerClientsIndexRoute extends Route {
+  beforeModel() {
+    return this.replaceWith('authenticated.partner.clients.overview');
+  }
+}
