+++ app/routes/authenticated/partner/clients/invitations.js
@@ -1,0 +1,13 @@
+import Route from '@ember/routing/route';
+import {inject as service} from '@ember/service';
+
+export default class AuthenticatedPartnerInvitationsRoute extends Route {
+  @service me;
+  @service partner;
+
+  beforeModel() {
+    if (!this.get('me.org.can_access_partner_dashboard') || !this.partner.access.invite_clients) {
+      this.transitionTo('authenticated.partner.clients');
+    }
+  }
+}
