+++ app/templates/authenticated/partner/clients.hbs
@@ -1,5 +1,4 @@
 {{page-title "Clients"}}
 
-<div class="padding-h-1 padding-b-2">
-  <Partner::ClientList></Partner::ClientList>
-</div>
+<Partner::ClientsNav></Partner::ClientsNav>
+{{outlet}}
