+++ app/templates/authenticated/partner/clients/invitations.hbs
@@ -1,0 +1,5 @@
+{{page-title "Invitations"}}
+
+<div class="margin-h-1 margin-b-4">
+  <Partner::InvitationList></Partner::InvitationList>
+</div>
