+++ app/templates/authenticated/partner/clients/overview.hbs
@@ -1,0 +1,3 @@
+{{page-title "Clients"}}
+
+<Partner::ClientList></Partner::ClientList>
