+++ app/templates/authenticated/partner/clients/registration-requests.hbs
@@ -1,0 +1,6 @@
+{{page-title "Registration Requests"}}
+
+<div class="margin-h-1 margin-b-4">
+  <Partner::RegistrationRequestPendingList></Partner::RegistrationRequestPendingList>
+  <Partner::RegistrationRequestRejectedList></Partner::RegistrationRequestRejectedList>
+</div>
