--- app/templates/authenticated/partner/invitations.hbs
@@ -1,3 +1,0 @@
-{{page-title "Invitations"}}
-<Partner::Invitations></Partner::Invitations>
-{{outlet}}
