--- app/templates/authenticated/partner/registration-requests.hbs
@@ -1,3 +1,0 @@
-{{page-title "Registration Requests"}}
-<Partner::RegistrationRequests></Partner::RegistrationRequests>
-{{outlet}}
