+++ tests/unit/routes/authenticated/partner/clients/index-test.js
@@ -1,0 +1,11 @@
+import { module, test } from 'qunit';
+import { setupTest } from 'ember-qunit';
+
+module('Unit | Route | authenticated/partner/clients/index', function(hooks) {
+  setupTest(hooks);
+
+  test('it exists', function(assert) {
+    let route = this.owner.lookup('route:authenticated/partner/clients/index');
+    assert.ok(route);
+  });
+});
